+++ custom-gcc/examples/redsquare/src/redsquare.c
@@ -1020,5 +1020,5 @@ void _start(void) {
     ReleaseMutex(mutex);
     CloseHandle(mutex);
 
-    ExitProcess((int32_t)(uintptr_t)message.wParam);
+    ExitProcess((uintptr_t)message.wParam);
 }
